pyproject.toml
[build-system]
requires = ['setuptools>=40.8.0']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
packages = ["fhirflat", "fhirflat.resources"]

[project]
name = "fhirflat"
version = "0.1.0"
description = "Flattened FHIR resources"
authors = [
    {name = "Pip Liggins", email = "philippa.liggins@dtc.ox.ac.uk"},
  {name = "Abhishek Dasgupta", email = "abhishek.dasgupta@dtc.ox.ac.uk"},
]
license = {file = "LICENSE"}
requires-python = ">=3.10"
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dependencies = [
    "fhir.resources==7.1.0",
    "numpy==1.26.4",
    "orjson==3.9.13",
    "pandas>=2.2.0",
    "pyarrow==15.0.0",
    "pydantic==2.6.1",
    "pydantic_core==2.16.2",
    "tzdata",
    "python-dateutil",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "pytest-unordered",
  "ruff",
  "tomli==2.*; python_version < '3.11'",
  "pre-commit"
]
docs = [
  "jupyter-book"
]

[project.urls]
Home = "https://github.com/globaldothealth/fhirflat"

[project.scripts]
  fhirflat = "fhirflat.__main__:main"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
pythonpath = "."

[tool.ruff]
exclude = [".venv", ".vscode", ".git",  "docs", "tests", "__init__.py"]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
extend-select = [
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "F",        # pyflakes
    "I",        # isort
    "C",        # flake8-comprehensions
    "B",        # flake8-bugbear
    "ARG",      # flake8-unused-arguments
    "RUF",      # Ruff-specific
    "YTT",      # flake8-2020
]
ignore = [
  "C901",       # function is too complex
  "C408",       # unnecessary `dict` call (rewrite as a literal)
]
